# Optimise prop query