--- a/automerge/src/op_tree.rs
+++ b/automerge/src/op_tree.rs
@@ -8,7 +8,7 @@ use std::{
 pub(crate) use crate::op_set::OpSetMetadata;
 use crate::{
     clock::Clock,
-    query::{self, Index, QueryResult, TreeQuery},
+    query::{self, Index, QueryResult, ReplaceArgs, TreeQuery},
 };
 use crate::{
     types::{ObjId, Op, OpId},
@@ -552,16 +552,24 @@ impl OpTreeNode {
     /// Update the operation at the given index using the provided function.
     ///
     /// This handles updating the indices after the update.
-    pub(crate) fn update<F>(&mut self, index: usize, f: F) -> (Op, &Op)
+    pub(crate) fn update<F>(&mut self, index: usize, f: F) -> ReplaceArgs
     where
         F: FnOnce(&mut Op),
     {
         if self.is_leaf() {
             let new_element = self.elements.get_mut(index).unwrap();
-            let old_element = new_element.clone();
+            let old_id = new_element.id;
+            let old_visible = new_element.visible();
             f(new_element);
-            self.index.replace(&old_element, new_element);
-            (old_element, new_element)
+            let replace_args = ReplaceArgs {
+                old_id,
+                new_id: new_element.id,
+                old_visible,
+                new_visible: new_element.visible(),
+                new_key: new_element.elemid_or_key(),
+            };
+            self.index.replace(&replace_args);
+            replace_args
         } else {
             let mut cumulative_len = 0;
             let len = self.len();
@@ -572,15 +580,23 @@ impl OpTreeNode {
                     }
                     Ordering::Equal => {
                         let new_element = self.elements.get_mut(child_index).unwrap();
-                        let old_element = new_element.clone();
+                        let old_id = new_element.id;
+                        let old_visible = new_element.visible();
                         f(new_element);
-                        self.index.replace(&old_element, new_element);
-                        return (old_element, new_element);
+                        let replace_args = ReplaceArgs {
+                            old_id,
+                            new_id: new_element.id,
+                            old_visible,
+                            new_visible: new_element.visible(),
+                            new_key: new_element.elemid_or_key(),
+                        };
+                        self.index.replace(&replace_args);
+                        return replace_args;
                     }
                     Ordering::Greater => {
-                        let (old_element, new_element) = child.update(index - cumulative_len, f);
-                        self.index.replace(&old_element, new_element);
-                        return (old_element, new_element);
+                        let replace_args = child.update(index - cumulative_len, f);
+                        self.index.replace(&replace_args);
+                        return replace_args;
                     }
                 }
             }