… automerge/Cargo.toml
… automerge/src/query/nth.rs
… automerge/src/query/seek_op_with_patch.rs