Results from critcmp using the new benchmarks:
```
group                       main                                    opt-prop
-----                       ----                                    --------
decreasing put 1000         1.00  1267.9±80.15µs        ? ?/sec     1.13  1438.7±39.22µs        ? ?/sec
decreasing put 10000        1.00     16.0±1.14ms        ? ?/sec     1.24     19.9±1.24ms        ? ?/sec
increasing put 1000         1.00  1460.2±103.95µs        ? ?/sec    1.20  1750.3±16.75µs        ? ?/sec
increasing put 10000        1.00     19.6±1.08ms        ? ?/sec     1.28     25.1±1.18ms        ? ?/sec
repeated increment 1000     9.17     11.7±1.14ms        ? ?/sec     1.00  1277.1±14.38µs        ? ?/sec
repeated increment 10000    98.99 1708.5±55.31ms        ? ?/sec     1.00     17.3±0.56ms        ? ?/sec
repeated put 1000           7.30     11.7±1.03ms        ? ?/sec     1.00  1605.7±43.10µs        ? ?/sec
repeated put 10000          79.08 1655.9±11.38ms        ? ?/sec     1.00     20.9±0.53ms        ? ?/sec
```

When there is a unique op for the key the old version is faster (currently, I have ideas to improve it later). But, when there are multiple updates to a key (`repeated *`) there is a massive speedup, basically by skipping non-visible ops and tracking visibility in the index, even for maps.

Fixes #349