@orionz 2022-04-26
Seems like this has a 25% speed penalty for normal puts and sets - is there any way to fix this without that baseline cost?
What if we kept `visible: Map<Key,Vec<Value>>` in the index?  Couldn't we dramatically reduce the search code for all query operations in non `_at` functions then?  Somewhat similar to the `OpRange<>` idea I was kicking around but would be super simple to implement and should give similar performance results..  All the nth queries would just be about converting pos to elemid and then looking in the index for the answer.