@jeffa5 2022-04-27
Conflicts fixed, I've also added a skip method to the searching which should bring back the performance for singles